requirements.txt
django==1.11
Pillow==4.3.0
django-wysiwyg-redactor
